Refactor package structure: reorganize schemas and controllers for improved clarity and consistency
