+++ src/wembed_core/controllers/dl_chunk_controller.py
@@ -6,7 +6,7 @@ from typing import List, Optional
 
 from wembed_core.database import DatabaseService
 from wembed_core.models.dl_doc.dl_chunks import DLChunks
-from wembed_core.schemas.dl_schemas import DLChunkSchema
+from wembed_core.schemas.dl_doc_schemas import DLChunkSchema
 
 
 class DLChunkController:
