+++ src/wembed_core/controllers/dl_input_controller.py
@@ -6,7 +6,7 @@ from typing import List, Optional
 
 from wembed_core.database import DatabaseService
 from wembed_core.models.dl_doc.dl_inputs import DLInputs
-from wembed_core.schemas.dl_schemas import DLInputSchema
+from wembed_core.schemas.dl_doc_schemas import DLInputSchema
 
 
 class DLInputController:
