+++ src/wembed_core/controllers/indexed_file_line_controller.py
@@ -6,7 +6,7 @@ from typing import List, Optional
 
 from wembed_core.database import DatabaseService
 from wembed_core.models.indexing.indexed_file_lines import IndexedFileLines
-from wembed_core.schemas.indexed_file_line_schema import IndexedFileLineSchema
+from wembed_core.schemas.indexing_schemas import IndexedFileLineSchema
 
 
 class IndexedFileLineController:
