+++ src/wembed_core/controllers/indexed_repo_controller.py
@@ -6,7 +6,7 @@ from typing import List, Optional
 
 from wembed_core.database import DatabaseService
 from wembed_core.models.indexing.indexed_repos import IndexedRepos
-from wembed_core.schemas.indexed_repo_schema import IndexedRepoSchema
+from wembed_core.schemas.indexing_schemas import IndexedRepoSchema
 
 
 class IndexedRepoController:
