+++ src/wembed_core/models/__init__.py
@@ -1,11 +1,2 @@
-from wembed_core.models.indexing.indexed_files import IndexedFiles  # noqa: F401
-from wembed_core.models.indexing.indexed_obsidian_vaults import (  # noqa: F401
-    IndexedObsidianVaults,
-)
-from wembed_core.models.indexing.indexed_repos import IndexedRepos  # noqa: F401
-from wembed_core.models.indexing.indexing_results import (  # noqa: F401
-    FileIndexingResults,
-)
-
-from .code_chunker import *  # noqa: F403, F401
-from .dl_doc import *  # noqa: F403, F401
+from .code_chunker import *  # noqa: F401, F403
+from .indexing import *  # noqa: F401, F403
