+++ src/wembed_core/schemas/__init__.py
@@ -1,10 +1,3 @@
-from .indexed_file_line_schema import IndexedFileLineSchema  # noqa: F401
-from .indexed_file_schema import IndexedFileSchema  # noqa: F401
-from .indexed_obsidian_vault_schema import (  # noqa: F401
-    IndexedObsidianVaultSchema,
-)
-from .indexed_repo_schema import IndexedRepoSchema  # noqa: F401
-from .indexing_result_schemas import (  # noqa: F401
-    IndexingResultSchema,
-    IndexingResultsListSchema,
-)
+from .code_chunker_schemas import *  # noqa: F403, F401
+from .dl_doc_schemas import *  # noqa: F403, F401
+from .indexing_schemas import *  # noqa: F403, F401
