+++ src/wembed_core/schemas/code_chunker_schemas.py
@@ -1,5 +1,5 @@
-import datetime
 import uuid
+from datetime import datetime
 from typing import List, Optional, Set
 
 from pydantic import BaseModel
@@ -155,3 +155,15 @@ class FunctionCall(BaseModel):
 
         arbitrary_types_allowed = True
         from_attributes = True
+
+
+__all__ = [
+    "CodeChunk",
+    "DependencyNode",
+    "FunctionCall",
+    "GitBranchSchema",
+    "GitCommitSchema",
+    "GitFileInfoSchema",
+    "ImportStatement",
+    "UsageNode",
+]
