+++ src/wembed_core/schemas/dl_doc_schemas.py
@@ -66,3 +66,6 @@ class DLInputSchema(BaseModel):
         """
 
         from_attributes = True
+
+
+__all__ = ["DLChunkSchema", "DLDocumentSchema", "DLInputSchema"]
